feat : 마이페이지 수정 기능 및 유저의 리뷰 조회 기능 추가

+++ src/main/java/yu/cse/odiga/auth/application/S3ProfileImageUploadService.java
@@ -2,7 +2,6 @@ package yu.cse.odiga.auth.application;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
@@ -24,25 +23,22 @@ public class S3ProfileImageUploadService {
     public void upload(MultipartFile multipartFile, User user) throws IOException {
         String originalFileName = multipartFile.getOriginalFilename();
         String uuid = UUID.randomUUID().toString();
-        String uniqueFileName = "profile/" + uuid + "_" + originalFileName.replaceAll("\\s", "_");
+        String uploadFileName = "profile/" + uuid + "_" + originalFileName.replaceAll("\\s", "_");
 
-        log.info("fileName: " + uniqueFileName);
+        log.info("업로드할 파일경로: " + uploadFileName);
 
         File uploadFile = s3Util.convert(multipartFile);
 
-        String uploadImageUrl = s3Util.putS3(uploadFile, uniqueFileName);
+        String uploadImageUrl = s3Util.putS3(uploadFile, uploadFileName);
         s3Util.removeNewFile(uploadFile);
 
         user.setProfileImageUrl(uploadImageUrl);
-
     }
 
     public void updateFile(MultipartFile newFile, String oldFileName, User user) throws IOException {
-        // 기존 파일 삭제
-        log.info("S3 oldFileName: " + oldFileName);
-        s3Util.deleteFile(oldFileName);
-        // 새 파일 업로드
+        if (oldFileName != null && !oldFileName.isEmpty()) {
+            s3Util.deleteFile(oldFileName);
+        }
         upload(newFile, user);
     }
-
 }
